requirements.txt
# Django
asgiref==3.4.1
Django==3.2.5
pytz==2021.1
sqlparse==0.4.1

# Django Rest Framework
djangorestframework==3.12.4

# Python Decouple
python-decouple==3.4
